+++ Source/WPEFramework/PluginServer.h
@@ -1689,7 +1689,7 @@ namespace WPEFramework {
                     }
                     else {
                         result = Core::ERROR_INVALID_SIGNATURE;
-                        uint32_t length = source.length();
+                        uint32_t length = static_cast<uint32_t>(source.length());
 
                         if ((callSign.length() == length) ||
                             ((callSign[length] == '.') && (index->second->HasVersionSupport(callSign.substr(length+1))))) {
@@ -2261,7 +2261,7 @@ namespace WPEFramework {
 							State(RAW, false);
 						}
 						if (Name().length() > (serviceHeader.length() + 1)) {
-							Properties(serviceHeader.length() + 1);
+							Properties(static_cast<uint32_t>(serviceHeader.length()) + 1);
 						}
 						// The state needs to be correct before we c
 						if (_service->Subscribe(*this) == false) {
@@ -2275,7 +2275,7 @@ namespace WPEFramework {
 					else {
 						const string& JSONRPCHeader(_parent._config.JSONRPCPrefix());
 						if (Name().length() > (JSONRPCHeader.length() + 1)) {
-							Properties(JSONRPCHeader.length() + 1);
+							Properties(static_cast<uint32_t>(JSONRPCHeader.length()) + 1);
 						}
 						State(JSONRPC, false);
